# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PokerHub API
  version: 3.0.0
paths:
...
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
        '429':
          description: Too many login attempts
  /auth/register:
    post:
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/logout:
    post:
      summary: Logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid token
  /auth/request-reset:
    post:
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestResetRequest'
      responses:
        '200':
          description: Reset requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/verify-reset-code:
    post:
      summary: Verify reset code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyResetCodeRequest'
      responses:
        '200':
          description: Code verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/reset-password:
    post:
      summary: Reset password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
  /admin/messages:
    get:
      summary: List user messages
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Messages list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMessagesResponse'
  /admin/messages/{id}/reply:
    post:
      summary: Reply to user message
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplyMessageRequest'
      responses:
        '200':
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /admin/messages/{id}/read:
    post:
      summary: Mark message as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMessage'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    AdminMessage:
      type: object
      properties:
        id:
          type: integer
        sender:
          type: string
        userId:
          type: string
        avatar:
          type: string
        subject:
          type: string
        preview:
          type: string
        content:
          type: string
        time:
          type: string
        read:
          type: boolean
      required:
        - id
        - sender
        - userId
        - avatar
        - subject
        - preview
        - content
        - time
        - read
    AdminMessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/AdminMessage'
      required:
        - messages
...
    ReplyMessageRequest:
      type: object
      properties:
        reply:
          type: string
      required:
        - reply
...
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    RequestResetRequest:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    VerifyResetCodeRequest:
      type: object
      properties:
        email:
          type: string
        code:
          type: string
      required:
        - email
        - code
    ResetPasswordRequest:
      type: object
      properties:
        email:
          type: string
        code:
          type: string
        password:
          type: string
      required:
        - email
        - code
        - password
...